Remove the validatable protocol
There were simpler ways to implement most
of its functionality. We can wait on having
an actual protocol until it's
a requested feature.

The validator.checker() function was beefed up
as a result, to provide greater control over the
error messages (since you can't fall back on
using the validatable protocol anymore). It was
also renamed to validator.expectTo().