Remove the "object special behavior"
Turns out, the special behavior I had seen with objects in union types,
that I had coded around, actually only happens under specific
circumstances - when an object literal
is being assigned to that type.
This isn't something we need to worry about in this project.

This "special behavior" was removed, and a lot of refactoring happened in the
rule-enforcing logic to simplify it.

Also, I removed the generic "your object matched multiple unions
at once, but none of them fully" error, and
replaced it for a more specific error listing
different things that went wrong.